pyproject.toml
[project]
name = "agent-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[mysql,postgresql,sql,sqlite]>=0.2.11",
    "python-a2a>=0.5.9",
]
